+++ src/components/hero/clock.tsx
@@ -1,10 +1,32 @@
-import { Clock as ClientClock } from "./clock.client";
+"use client";
+import { useState, useEffect } from "react";
 
-export const Clock = ({ className }: { className?: string }) => {
-  const time = new Date();
-  const newYorkTime = time.toLocaleTimeString("en-US", {
-    timeZone: "America/New_York",
-  });
+export const Clock = () => {
+  const [time, setTime] = useState<string>();
+  const ONE_SECOND = 1000;
 
-  return <ClientClock className={className} initialTime={newYorkTime} />;
+  useEffect(() => {
+    const updateTime = () => {
+      const newTime = new Date().toLocaleTimeString("en-US", {
+        timeZone: "America/New_York",
+      });
+      setTime(newTime);
+    };
+
+    const interval = setInterval(updateTime, ONE_SECOND);
+    return () => clearInterval(interval);
+  }, []);
+
+  return (
+    <p className="text-sm mb-8">
+      EDT 🗽 /{" "}
+      {!!time ? (
+        <time dateTime={time} suppressHydrationWarning>
+          {time}
+        </time>
+      ) : (
+        <span className="ml-1 rounded-sm inline-block w-16 h-3 animate-pulse bg-slate-500/40"></span>
+      )}
+    </p>
+  );
 };
